generaliser visning av redaksjonelt innhold
Benytter samme komponenter for skjema og visning på tvers
av tiltakstyper/avtaler/gjennomføringer
