--- frontend/mr-admin-flate/src/components/redaksjoneltInnhold/RedaksjoneltInnholdPreview.tsx
@@ -1,395 +1,0 @@
-import { useTiltakstypeFaneinnhold } from "@/api/gjennomforing/useTiltakstypeFaneinnhold";
-import { Alert, BodyLong, Heading, HStack, Label, Link, VStack } from "@navikt/ds-react";
-import { LokalInformasjonContainer, PortableText } from "@mr/frontend-common";
-import { Suspense, useState } from "react";
-import { Laster } from "../laster/Laster";
-import { LenkerList } from "../lenker/LenkerList";
-import { RedaksjoneltInnholdContainer } from "@/components/redaksjoneltInnhold/RedaksjoneltInnholdContainer";
-import { TwoColumnGrid } from "@/layouts/TwoColumGrid";
-import { useGetAvtaleIdFromUrlOrThrow } from "@/hooks/useGetAvtaleIdFromUrl";
-import { useAvtale } from "@/api/avtaler/useAvtale";
-import { Bolk } from "../detaljside/Bolk";
-import { gjennomforingTekster } from "../ledetekster/gjennomforingLedetekster";
-import { Kontaktperson } from "@/pages/gjennomforing/Kontaktperson";
-import { CaretDownFillIcon, CaretUpFillIcon } from "@navikt/aksel-icons";
-import {
-  Faneinnhold,
-  GjennomforingKontaktpersonDto,
-  GjennomforingTiltakstype,
-  Kontorstruktur,
-  Regelverklenke,
-} from "@tiltaksadministrasjon/api-client";
-import { MetadataVStack } from "@mr/frontend-common/components/datadriven/Metadata";
-
-interface RedaksjoneltInnholdPreviewProps {
-  tiltakstype: GjennomforingTiltakstype;
-  beskrivelse: string | null;
-  faneinnhold: Faneinnhold | null;
-  kontorstruktur: Kontorstruktur[];
-  kontaktpersoner: GjennomforingKontaktpersonDto[];
-}
-
-export function RedaksjoneltInnholdPreview() {
-  const avtaleId = useGetAvtaleIdFromUrlOrThrow();
-  const { data: avtale } = useAvtale(avtaleId);
-  const { tiltakstype, beskrivelse, faneinnhold, kontorstruktur } = avtale;
-
-  return (
-    <Suspense fallback={<Laster tekst="Laster innhold" />}>
-      <RedaksjoneltInnhold
-        tiltakstype={tiltakstype}
-        kontorstruktur={kontorstruktur}
-        beskrivelse={beskrivelse}
-        faneinnhold={faneinnhold}
-        kontaktpersoner={[]}
-      />
-    </Suspense>
-  );
-}
-
-export function RedaksjoneltInnhold(props: RedaksjoneltInnholdPreviewProps) {
-  const { tiltakstype, kontaktpersoner, beskrivelse, faneinnhold, kontorstruktur } = props;
-
-  const { data: tiltakstypeSanityData } = useTiltakstypeFaneinnhold(tiltakstype.id);
-  return (
-    <TwoColumnGrid separator>
-      <RedaksjoneltInnholdContainer>
-        {tiltakstypeSanityData.beskrivelse && (
-          <>
-            <Heading size="medium">Generell informasjon</Heading>
-            <BodyLong size="large" style={{ whiteSpace: "pre-wrap" }}>
-              {tiltakstypeSanityData.beskrivelse}
-            </BodyLong>
-          </>
-        )}
-        {beskrivelse && (
-          <LokalInformasjonContainer>
-            <BodyLong style={{ whiteSpace: "pre-wrap" }} textColor="subtle" size="medium">
-              {beskrivelse}
-            </BodyLong>
-          </LokalInformasjonContainer>
-        )}
-        {someValuesExists([
-          faneinnhold?.forHvem,
-          faneinnhold?.forHvemInfoboks,
-          tiltakstypeSanityData.faneinnhold?.forHvem,
-          tiltakstypeSanityData.faneinnhold?.forHvemInfoboks,
-        ]) ? (
-          <>
-            <Heading size="medium">For hvem</Heading>
-            <DetaljerFane
-              gjennomforing={faneinnhold?.forHvem}
-              gjennomforingAlert={faneinnhold?.forHvemInfoboks}
-              tiltakstype={tiltakstypeSanityData.faneinnhold?.forHvem}
-              tiltakstypeAlert={tiltakstypeSanityData.faneinnhold?.forHvemInfoboks}
-            />
-          </>
-        ) : null}
-
-        {someValuesExists([
-          faneinnhold?.detaljerOgInnhold,
-          faneinnhold?.detaljerOgInnholdInfoboks,
-          tiltakstypeSanityData.faneinnhold?.detaljerOgInnhold,
-          tiltakstypeSanityData.faneinnhold?.detaljerOgInnholdInfoboks,
-        ]) ? (
-          <>
-            <Heading size="medium">Detaljer og innhold</Heading>
-            <DetaljerFane
-              gjennomforing={faneinnhold?.detaljerOgInnhold}
-              gjennomforingAlert={faneinnhold?.detaljerOgInnholdInfoboks}
-              tiltakstype={tiltakstypeSanityData.faneinnhold?.detaljerOgInnhold}
-              tiltakstypeAlert={tiltakstypeSanityData.faneinnhold?.detaljerOgInnholdInfoboks}
-            />
-          </>
-        ) : null}
-
-        {someValuesExists([
-          faneinnhold?.pameldingOgVarighet,
-          faneinnhold?.pameldingOgVarighetInfoboks,
-          tiltakstypeSanityData.faneinnhold?.pameldingOgVarighet,
-          tiltakstypeSanityData.faneinnhold?.pameldingOgVarighetInfoboks,
-        ]) ? (
-          <>
-            <Heading size="medium">Påmelding og varighet</Heading>
-            <DetaljerFane
-              gjennomforing={faneinnhold?.pameldingOgVarighet}
-              gjennomforingAlert={faneinnhold?.pameldingOgVarighetInfoboks}
-              tiltakstype={tiltakstypeSanityData.faneinnhold?.pameldingOgVarighet}
-              tiltakstypeAlert={tiltakstypeSanityData.faneinnhold?.pameldingOgVarighetInfoboks}
-            />
-          </>
-        ) : null}
-
-        {someValuesExists([faneinnhold?.kontaktinfo, faneinnhold?.kontaktinfoInfoboks]) ? (
-          <>
-            <Heading size="medium">Kontaktinfo</Heading>
-            <DetaljerFane
-              gjennomforing={faneinnhold?.kontaktinfo}
-              gjennomforingAlert={faneinnhold?.kontaktinfoInfoboks}
-            />
-          </>
-        ) : null}
-
-        {someValuesExists([faneinnhold?.lenker]) ? (
-          <div className="prose">
-            <Heading size="medium">Lenker</Heading>
-            <LenkerList lenker={faneinnhold?.lenker || []} />
-          </div>
-        ) : null}
-
-        {someValuesExists([faneinnhold?.delMedBruker, tiltakstypeSanityData.delingMedBruker]) ? (
-          <>
-            <Heading size="medium">Del med bruker</Heading>
-            <BodyLong as="div" size="small" className="prose">
-              {faneinnhold?.delMedBruker ?? tiltakstypeSanityData.delingMedBruker}
-            </BodyLong>
-          </>
-        ) : null}
-      </RedaksjoneltInnholdContainer>
-      <RedaksjoneltInnholdContainer>
-        <Bolk aria-label={gjennomforingTekster.tilgjengeligIModiaLabel}>
-          <MetadataVStack
-            label={gjennomforingTekster.tilgjengeligIModiaLabel}
-            value={<RegionOgUnderenheter kontorstruktur={kontorstruktur} />}
-          />
-        </Bolk>
-        {kontaktpersoner.length > 0 && (
-          <Bolk>
-            <MetadataVStack
-              label={gjennomforingTekster.kontaktpersonNav.mainLabel}
-              value={
-                <VStack gap="space-8">
-                  {kontaktpersoner.map((kp, index) => (
-                    <Kontaktperson key={index} kontaktperson={kp} />
-                  ))}
-                </VStack>
-              }
-            />
-          </Bolk>
-        )}
-      </RedaksjoneltInnholdContainer>
-    </TwoColumnGrid>
-  );
-}
-
-interface TiltakstypeFaneinnholdContentProps {
-  beskrivelse: string | null;
-  faneinnhold: Faneinnhold | null;
-  regelverklenker?: Regelverklenke[];
-  kanKombineresMed?: string[];
-}
-
-export function TiltakstypeFaneinnholdContent({
-  beskrivelse,
-  faneinnhold,
-  regelverklenker,
-  kanKombineresMed,
-}: TiltakstypeFaneinnholdContentProps) {
-  return (
-    <RedaksjoneltInnholdContainer>
-      {beskrivelse && (
-        <>
-          <Heading size="medium">Generell informasjon</Heading>
-          <BodyLong size="large" style={{ whiteSpace: "pre-wrap" }}>
-            {beskrivelse}
-          </BodyLong>
-        </>
-      )}
-
-      {someValuesExists([faneinnhold?.forHvem, faneinnhold?.forHvemInfoboks]) ? (
-        <>
-          <Heading size="medium">For hvem</Heading>
-          <DetaljerFane
-            tiltakstype={faneinnhold?.forHvem}
-            tiltakstypeAlert={faneinnhold?.forHvemInfoboks}
-          />
-        </>
-      ) : null}
-
-      {someValuesExists([
-        faneinnhold?.detaljerOgInnhold,
-        faneinnhold?.detaljerOgInnholdInfoboks,
-      ]) ? (
-        <>
-          <Heading size="medium">Detaljer og innhold</Heading>
-          <DetaljerFane
-            tiltakstype={faneinnhold?.detaljerOgInnhold}
-            tiltakstypeAlert={faneinnhold?.detaljerOgInnholdInfoboks}
-          />
-        </>
-      ) : null}
-
-      {someValuesExists([
-        faneinnhold?.pameldingOgVarighet,
-        faneinnhold?.pameldingOgVarighetInfoboks,
-      ]) ? (
-        <>
-          <Heading size="medium">Påmelding og varighet</Heading>
-          <DetaljerFane
-            tiltakstype={faneinnhold?.pameldingOgVarighet}
-            tiltakstypeAlert={faneinnhold?.pameldingOgVarighetInfoboks}
-          />
-        </>
-      ) : null}
-
-      {someValuesExists([faneinnhold?.kontaktinfo, faneinnhold?.kontaktinfoInfoboks]) ? (
-        <>
-          <Heading size="medium">Kontaktinfo</Heading>
-          <DetaljerFane
-            tiltakstype={faneinnhold?.kontaktinfo}
-            tiltakstypeAlert={faneinnhold?.kontaktinfoInfoboks}
-          />
-        </>
-      ) : null}
-
-      {someValuesExists([faneinnhold?.lenker]) ? (
-        <div className="prose">
-          <Heading size="medium">Lenker</Heading>
-          <LenkerList lenker={faneinnhold?.lenker || []} />
-        </div>
-      ) : null}
-
-      {faneinnhold?.delMedBruker ? (
-        <>
-          <Heading size="medium">Del med bruker</Heading>
-          <BodyLong as="div" size="small" className="prose">
-            {faneinnhold.delMedBruker}
-          </BodyLong>
-        </>
-      ) : null}
-
-      {regelverklenker && regelverklenker.length > 0 && (
-        <>
-          <Heading size="medium">Regelverk</Heading>
-          <VStack gap="space-2">
-            {regelverklenker.map((lenke) => (
-              <HStack key={lenke.regelverkUrl} gap="space-4" align="center">
-                <Link href={lenke.regelverkUrl} target="_blank">
-                  {lenke.regelverkLenkeNavn ?? lenke.regelverkUrl}
-                </Link>
-                {lenke.beskrivelse && (
-                  <Label size="small" as="span">
-                    {lenke.beskrivelse}
-                  </Label>
-                )}
-              </HStack>
-            ))}
-          </VStack>
-        </>
-      )}
-
-      {kanKombineresMed && kanKombineresMed.length > 0 && (
-        <>
-          <Heading size="medium">Kan kombineres med</Heading>
-          <VStack gap="space-2">
-            {kanKombineresMed.map((navn) => (
-              <BodyLong key={navn}>{navn}</BodyLong>
-            ))}
-          </VStack>
-        </>
-      )}
-    </RedaksjoneltInnholdContainer>
-  );
-}
-
-function someValuesExists(params: any[]): boolean {
-  return params.some((p) => !!p);
-}
-
-interface DetaljerFaneProps {
-  gjennomforingAlert?: string | null;
-  tiltakstypeAlert?: string | null;
-  gjennomforing?: any;
-  tiltakstype?: any;
-}
-
-const DetaljerFane = ({
-  gjennomforingAlert,
-  tiltakstypeAlert,
-  gjennomforing,
-  tiltakstype,
-}: DetaljerFaneProps) => {
-  if (!gjennomforingAlert && !tiltakstypeAlert && !gjennomforing && !tiltakstype) {
-    return <></>;
-  }
-
-  return (
-    <div>
-      {tiltakstype && (
-        <>
-          {tiltakstypeAlert && (
-            <Alert style={{ whiteSpace: "pre-wrap" }} variant="info">
-              {tiltakstypeAlert}
-            </Alert>
-          )}
-          <BodyLong as="div" size="small">
-            <PortableText value={tiltakstype} />
-          </BodyLong>
-        </>
-      )}
-      {(gjennomforing || gjennomforingAlert) && (
-        <LokalInformasjonContainer>
-          <Heading level="2" size="small" spacing className="mt-0">
-            Lokal Informasjon
-          </Heading>
-          {gjennomforingAlert && (
-            <Alert style={{ whiteSpace: "pre-wrap" }} variant="info">
-              {gjennomforingAlert}
-            </Alert>
-          )}
-          <BodyLong as="div" size="small">
-            <PortableText value={gjennomforing} />
-          </BodyLong>
-        </LokalInformasjonContainer>
-      )}
-    </div>
-  );
-};
-
-function RegionOgUnderenheter({ kontorstruktur }: { kontorstruktur: Kontorstruktur[] }) {
-  const [openRegions, setOpenRegions] = useState<string[]>([]);
-
-  const toggleRegion = (enhetsnummer: string) => {
-    setOpenRegions((prev) =>
-      prev.includes(enhetsnummer)
-        ? prev.filter((num) => num !== enhetsnummer)
-        : [...prev, enhetsnummer],
-    );
-  };
-
-  function isRegionOpen(enhetsnummer: string) {
-    return openRegions.includes(enhetsnummer);
-  }
-
-  return (
-    <ul>
-      {kontorstruktur.map((kontor) => {
-        return (
-          <li className="font-ax-bold my-2 ml-3" key={kontor.region.enhetsnummer}>
-            <button
-              className="hover:cursor-pointer flex"
-              onClick={() => toggleRegion(kontor.region.enhetsnummer)}
-              title={`${kontor.region.navn} (${kontor.kontorer.length} kontorer)`}
-            >
-              {kontor.region.navn} ({kontor.kontorer.length || 0})
-              {isRegionOpen(kontor.region.enhetsnummer) ? (
-                <CaretUpFillIcon className="text-xl" />
-              ) : (
-                <CaretDownFillIcon className="text-xl" />
-              )}
-            </button>
-            {isRegionOpen(kontor.region.enhetsnummer) && (
-              <ul className="list-disc ml-5">
-                {kontor.kontorer.map((kontor) => (
-                  <li className="ml-5 font-thin" key={kontor.enhetsnummer}>
-                    {kontor.navn}
-                  </li>
-                ))}
-              </ul>
-            )}
-          </li>
-        );
-      })}
-    </ul>
-  );
-}
